update: visualization (split/single)
