+++ static_report/src/components/shared/ColumnCard/index.tsx
@@ -2,13 +2,11 @@ import { Flex, Text } from '@chakra-ui/react';
 import { Link, useLocation } from 'wouter';
 import { ColumnSchema } from '../../../sdlc/single-report-schema';
 import { ZColSchema } from '../../../types';
+import { getDataChart } from '../../../utils/charts';
 import {
   checkColumnCategorical,
   getChartKindByColumnType,
 } from '../../../utils/transformers';
-import { BooleanPieChart } from '../Charts/BooleanPieChart';
-import { CategoricalBarChart } from '../Charts/CategoricalBarChart';
-import { HistogramChart } from '../Charts/HistogramChart';
 import { ColumnCardBodyContainer } from './ColumnCardBodyContainer';
 import { ColumnCardDataVisualContainer } from './ColumnCardDataVisualContainer';
 import { ColumnCardHeader } from './ColumnCardHeader';
@@ -18,7 +16,6 @@ import { ColumnTypeDetailDatetime } from './ColumnTypeDetail/ColumnTypeDetailDat
 import { ColumnTypeDetailNumeric } from './ColumnTypeDetail/ColumnTypeDetailNumeric';
 import { ColumnTypeDetailOther } from './ColumnTypeDetail/ColumnTypeDetailOther';
 import { ColumnTypeDetailText } from './ColumnTypeDetail/ColumnTypeDetailText';
-import { FALSES, INVALIDS, NULLS, TRUES } from './ColumnTypeDetail/constants';
 
 interface Props {
   columnDatum: ColumnSchema;
@@ -55,70 +52,6 @@ export function ColumnCard({ columnDatum }: Props) {
           </Text>
         </Link>
       </Flex>
-    </Flex>
-  );
-}
-
-/**
- * Handles logic for rendering the right charts
- * @param columnDatum
- * @param baseColumnRef an optional column reference for comparing `target` against `base` columns, to ensure that the chart kind is consistent across comparisons members
- * @returns *Chart Component
- */
-export function getDataChart(
-  columnDatum: ColumnSchema,
-  baseColumnRef?: ColumnSchema,
-) {
-  const {
-    total,
-    name,
-    type,
-    schema_type,
-    histogram,
-    topk,
-    trues,
-    falses,
-    nulls,
-    invalids,
-    valids,
-    min,
-    max,
-  } = columnDatum;
-
-  const hasSameTypeName =
-    type === baseColumnRef?.type && name === baseColumnRef?.name;
-
-  //FIXME: Override flag?
-  const chartKind = getChartKindByColumnType(
-    hasSameTypeName ? baseColumnRef : columnDatum,
-  );
-
-  //TopK dataset
-  if (chartKind === 'topk' && topk) {
-    return <CategoricalBarChart data={topk} total={total || 0} />;
-  }
-  //histogram dataset
-  if (chartKind === 'histogram' && histogram) {
-    return <HistogramChart data={{ histogram, min, max, type, total }} />;
-  }
-  //pie dataset
-  if (chartKind === 'pie') {
-    const counts = [trues, falses, nulls, invalids].map((v) => (v ? v : 0));
-    const labels = [TRUES, FALSES, NULLS, INVALIDS].map(
-      (v) => v.charAt(0) + v.slice(1).toLowerCase(),
-    );
-    const ratios = counts.map((v) => v / Number(total));
-    return <BooleanPieChart data={{ counts, labels, ratios }} />;
-  }
-
-  const noRenderMessage = Boolean(valids)
-    ? `Chart rendering unavailable for (type: ${schema_type})`
-    : `There are insufficient valid data points in this dataset`;
-  return (
-    <Flex h={230} alignItems={'center'} w={'100%'}>
-      <Text textAlign={'center'} w={'inherit'}>
-        {noRenderMessage}
-      </Text>
     </Flex>
   );
 }
